requirements.txt
Django>=4.0.2,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
requests>=2.26.0,<2.27
langchain       # no version pin here
langchain-ollama    # no version pins, or pin one if needed: e.g. langchain-ollama==0.1.0
pandas>=1.4.2,<1.5
numpy>=1.21.5,<1.22
python-dotenv>=0.19.2,<0.20
openpyxl>=3.0.0,<3.1